Break out FamilyDataContext and useFamilyData hook
