+++ client/src/contexts/FamilyDataContext.tsx
@@ -1,30 +1,7 @@
-import React, {
-  createContext,
-  useContext,
-  useCallback,
-  useMemo,
-  type ReactNode,
-} from 'react';
+import React, { useMemo, type ReactNode } from 'react';
 import { useGedcomData } from '../hooks/useGedcomData';
 import type { GedcomDataWithMetadata } from '../../../shared/types';
-
-// Status discriminated union as mentioned in the issue requirements
-type FamilyDataStatus = 
-  | { type: 'idle'; data: null; error: null }
-  | { type: 'loading'; data: null; error: null }
-  | { type: 'success'; data: GedcomDataWithMetadata; error: null }
-  | { type: 'error'; data: null; error: string };
-
-interface FamilyDataContextValue {
-  status: FamilyDataStatus['type'];
-  data: GedcomDataWithMetadata | null;
-  error: string | null;
-  refetch: () => void;
-}
-
-const FamilyDataContext = createContext<FamilyDataContextValue | undefined>(
-  undefined,
-);
+import { FamilyDataContext, type FamilyDataContextValue } from './familyDataContextDefinition';
 
 interface FamilyDataProviderProps {
   jsonFile: string;
@@ -46,7 +23,7 @@ export function FamilyDataProvider({
   });
 
   // Determine status based on loading/error/data state
-  const status: FamilyDataStatus['type'] = useMemo(() => {
+  const status: FamilyDataContextValue['status'] = useMemo(() => {
     if (loading) return 'loading';
     if (error) return 'error';
     if (data) return 'success';
@@ -67,42 +44,5 @@ export function FamilyDataProvider({
     <FamilyDataContext.Provider value={contextValue}>
       {children}
     </FamilyDataContext.Provider>
-  );
-}
-
-export function useFamilyData(): FamilyDataContextValue {
-  const context = useContext(FamilyDataContext);
-  if (context === undefined) {
-    throw new Error('useFamilyData must be used within a FamilyDataProvider');
-  }
-  return context;
-}
-
-// Provider component keyed by jsonFile to prevent unnecessary re-renders
-const providers = new Map<string, React.ComponentType<{ children: ReactNode }>>();
-
-interface FamilyDataProviderWrapperProps {
-  jsonFile: string;
-  children: ReactNode;
-  onDataLoaded?: (data: GedcomDataWithMetadata) => void;
-  onError?: (error: string) => void;
-}
-
-export function createKeyedFamilyDataProvider({
-  jsonFile,
-  children,
-  onDataLoaded,
-  onError,
-}: FamilyDataProviderWrapperProps): React.ReactElement {
-  // Create a keyed provider to ensure proper cache invalidation
-  return (
-    <FamilyDataProvider
-      key={jsonFile}
-      jsonFile={jsonFile}
-      onDataLoaded={onDataLoaded}
-      onError={onError}
-    >
-      {children}
-    </FamilyDataProvider>
   );
 }
